Fixed simulation.c by properly wrapping around the sides of the grid (modulo operation); also fixed order of operations to take the last generation into account for species count. Serial implementation now fully works as intended.
Now long data type is used for max_counts.

+++ serial/include/grid.h
@@ -4,7 +4,7 @@
 
 // Function declarations
 char ***gen_initial_grid(long long N, float density, int input_seed);
-void simulation(char ***grid, int N, int *max_counts, int *max_generations, int num_generations);
-void print_result(int max_counts[N_SPECIES + 1], int max_generations[N_SPECIES + 1]);
+void simulation(char ***grid, int N, long *max_counts, int *max_generations, int num_generations);
+void print_result(long max_counts[N_SPECIES + 1], int max_generations[N_SPECIES + 1]);
 
 void print_grid(char ***grid, int N);
